affinity.spline.coefs {gcrma}	R Documentation

Spline coefficients for estimation of affinity from probe sequence

Description

Spline coefficients for estimation of affinity from probe sequence

Usage

data(affinity.spline.coefs)

See Also

compute.affinities

[Package gcrma version 2.12.1 Index]