bg.adjust.affinities {gcrma}	R Documentation

Background adjustment with sequence information (internal function)

Description

An internal function to be used by gcrma.

Usage


bg.adjust.fullmodel(pms,mms,ncs=NULL,apm,amm,anc=NULL,index.affinities,k=k,rho=.7,fast)
bg.adjust.affinities(pms,ncs,apm,anc,index.affinities,k=k,fast=FALSE,nomm=FALSE)

Arguments

pms 	 PM intensities after optical background correction, before non-specific-binding correction.

mms 	 MM intensities after optical background correction, before non-specific-binding correction.

ncs 	 Negative control probe intensities after optical background correction, before non-specific-binding correction. If ncs=NULL, the MM probes are considered the negative control probes.

index.affinities 	 The index of pms with known sequences. (For some types of arrays the sequences of a small subset of probes are not provided by Affymetrix.)

apm 	 Probe affinities for PM probes with known sequences.

amm 	 Probe affinities for MM probes with known sequences.

anc 	 Probe affinities for Negative control probes with known sequences. This is ignored when ncs=NULL.

rho 	 correlation coefficient of log background intensity in a pair of pm/mm probes. Default=.7

k 	 A tuning parameter. See details.

fast 	 Logical value. If TRUE a faster add-hoc algorithm is used.

nomm

Details

Assumes PM=background1+signal,mm=background2, (log(background1),log(background2))' follow bivariate normal distribution, signal distribution follows power law. bg.parameters.gcrma and sg.parameters.gcrma provide adhoc estimates of the parameters.

the original gcrma uses an empirical Bayes estimate. this requires a complicated numerical integration. An add-hoc method tries to imitate the empirical Bayes estimate with a PM-B but values of PM-B<k going to k. This can be thought as a shrunken MVUE. For more details see Wu et al. (2003).

Value

a vector of same length as x.

Author(s)

Rafeal Irizarry, Zhijin(Jean) Wu

See Also

gcrma

[Package gcrma version 2.12.1 Index]