bg.adjust.gcrma {gcrma}	R Documentation

GCRMA background adjust (internal function)

Description

This function performs background adjustment (optical noise and non-specific binding on an AffyBatch project and returns an AffyBatch object in which the PM intensities are adjusted.

Usage


bg.adjust.gcrma(object,affinity.info=NULL,
      affinity.source=c("reference","local"),
      NCprobe=NULL,
      type=c("fullmodel","affinities","mm","constant"),
      k=6*fast+0.5*(1-fast),stretch=1.15*fast+1*(1-fast),correction=1,
      GSB.adjust=TRUE,
      rho=.7,optical.correct=TRUE,verbose=TRUE,fast=TRUE)

Arguments

object 	 an AffyBatch

affinity.info 	 NULL or an AffyBatch containing the affinities in the exprs slot. This object can be created using the function compute.affinities.

affinity.source 	 reference: use the package internal Non-specific binding data or local: use the experimental data in object. If local is chosen, either MM probes or a user-defined list of probes (see NCprobes) are used to estimate affinities.

NCprobe

type 	 "fullmodel" for sequence and MM model. "affinities" for sequence information only. "mm" for using MM without sequence information.

k 	 A tuning factor.

stretch

correction 	 .

GSB.adjust 	 Logical value. If TRUE, probe effects in specific binding will be adjusted.

rho 	 correlation coefficient of log background intensity in a pair of pm/mm probes. Default=.7

optical.correct 	 Logical value. If TRUE, optical background correction is performed.

verbose 	 Logical value. If TRUE messages about the progress of the function is printed.

fast 	 Logical value. If TRUE a faster ad hoc algorithm is used.

Details

The returned value is an AffyBatch object, in which the PM probe intensities have been background adjusted. The rest is left the same as the starting AffyBatch object.

The tunning factor k will have different meainngs if one uses the fast (ad hoc) algorithm or the empirical bayes approach. See Wu et al. (2003)

Value

An AffyBatch.

Author(s)

Rafeal Irizarry

Examples


 if(require(affydata) & require(hgu95av2probe) & require(hgu95av2cdf)){
          data(Dilution)
          ai <- compute.affinities(cdfName(Dilution))
          Dil.adj<-bg.adjust.gcrma(Dilution,affinity.info=ai,type="affinities")
     }

[Package gcrma version 2.12.1 Index]