compute.affinities {gcrma}	R Documentation

Probe Affinity computation

Description

An internal function to calculate probe affinities from their sequences.

Usage


compute.affinities(cdfname,verbose=TRUE)

check.probes(probepackage,cdfname)

Arguments

cdfname 	 Object of class character representing the name of CDF file associated with the arrays in the AffyBatch.

probepackage 	 character representing the name of the package with the probe sequence information.

verbose 	 Logical value. If TRUE messages about the progress of the function is printed.

Details

The affinity of a probe is described as the sum of position-dependent base affinities. Each base at each position contributes to the total affinity of a probe in an additive fashion. For a given type of base, the positional effect is modeled as a spline function with 5 degrees of freedom.

check.probes makes sure things are matching as they should.

Value

compute.affinities returns an AffyBatch with the affinities for PM probes in the pm locations and the affinities for MM probes in the mm locations. NA will be added for probes with no sequence information.

Author(s)

Rafeal Irizarry

References

Hekstra, D., Taussig, A. R., Magnasco, M., and Naef, F. (2003) Absolute mRNA concentrations from sequence-specific calibration of oligonucleotide array. Nucleic Acids Research, 31. 1962-1968.

See Also

gcrma,affinity.spline.coefs

[Package gcrma version 2.12.1 Index]