fast.bkg {gcrma}	R Documentation

Internal functions for justGCRMA

Description

These are internal functions for justGCRMA that are called based on memory or speed constraints.

Usage


fast.bkg(filenames, pm.affinities, mm.affinities, index.affinities,
type, minimum, optical.correct, verbose, k, rho, correction, stretch,
fast, cdfname, read.verbose)
mem.bkg(filenames, pm.affinities, mm.affinities, index.affinities, type,
minimum, optical.correct, verbose, k, rho, correction, stretch, fast,
cdfname, read.verbose)

Arguments

filenames 	 A list of cel files.

pm.affinities 	 Values passed from compute.affinities.

mm.affinities 	 Values passed from compute.affinities.

index.affinities 	 Values passed from compute.affinities.

type 	 "fullmodel" for sequence and MM model. "affinities" for sequence information only. "mm" for using MM without sequence information.

minimum 	 A minimum value to be used for optical.correct.

optical.correct 	 Logical value. If TRUE, optical background correction is performed.

verbose 	 Logical value. If TRUE, messages about the progress of the function are printed.

k 	 A tuning factor.

rho 	 correlation coefficient of log background intensity in a pair of pm/mm probes. Default=.7

correction

stretch

fast 	 Logical value. If TRUE, then a faster ad hoc algorithm is used.

cdfname 	 Used to specify the name of an alternative cdf package. If set to NULL, the usual cdf package based on Affymetrix' mappings will be used. Note that the name should not include the 'cdf' on the end, and that the corresponding probe package is also required to be installed. If either package is missing an error will result.

read.verbose 	 Logical value. If TRUE, a message is returned as each celfile is read in.

Details

Note that this expression measure is given to you in log base 2 scale. This differs from most of the other expression measure methods.

The tuning factor 'k' will have different meanings if one uses the fast (add-hoc) algorithm or the empirical Bayes approach. See Wu et al. (2003)

Value

An ExpressionSet.

Author(s)

James W. MacDonald <jmacdon@med.umich.edu>

See Also

gcrma

[Package gcrma version 2.12.1 Index]