gcrma {gcrma}	R Documentation

Robust Multi-Array expression measure using sequence information

Description

This function converts an AffyBatch into an ExpressionSet using the robust multi-array average (RMA) expression measure with help of probe sequence.

Usage


gcrma(object,affinity.info=NULL,
      affinity.source=c("reference","local"),NCprobe=NULL,
      type=c("fullmodel","affinities","mm","constant"),
      k=6*fast+0.5*(1-fast),stretch=1.15*fast+1*(1-fast),correction=1,
      GSB.adjust=TRUE,
      rho=.7,optical.correct=TRUE,verbose=TRUE,fast=TRUE,
      subset=NULL,normalize=TRUE,...)

Arguments

object 	 an AffyBatch

affinity.info 	 NULL or an AffyBatch containing the affinities in the exprs slot. This object can be created using the function compute.affinities.

affinity.source 	 reference: use the package internal Non-specific binding data or local: use the experimental data in object. If local is chosen, either MM probes or a user-defined list of probes (see NCprobes) are used to estimate affinities.

NCprobe 	 Index of negative control probes. When set as NULL,the MM probes will be used. These probes are used to estimate parameters of non-specific binding on each array. These will be also used to estimate probe affinity profiles when affinity.info is not provided.

type 	 "fullmodel" for sequence and MM model. "affinities" for sequence information only. "mm" for using MM without sequence information.

k 	 A tuning factor.

stretch

correction 	 .

GSB.adjust 	 Logical value. If TRUE, probe effects in specific binding will be adjusted.

rho 	 correlation coefficient of log background intensity in a pair of pm/mm probes. Default=.7

optical.correct 	 Logical value. If TRUE, optical background correction is performed.

verbose 	 Logical value. If TRUE messages about the progress of the function is printed.

fast 	 Logical value. If TRUE a faster ad hoc algorithm is used.

subset 	 a character vector with the the names of the probesets to be used in expression calculation.

normalize 	 logical value. If 'TRUE' normalize data using quantile normalization.

... 	 further arguments to be passed (not currently implemented - stub for future use).

Details

Note that this expression measure is given to you in log base 2 scale. This differs from most of the other expression measure methods.

The tuning factor k will have different meanings if one uses the fast (add-hoc) algorithm or the empirical Bayes approach. See Wu et al. (2003)

Value

An ExpressionSet.

Author(s)

Rafeal Irizarry

Examples


if(require(affydata) & require(hgu95av2probe) & require(hgu95av2cdf)){
     data(Dilution)
     ai <- compute.affinities(cdfName(Dilution))
     Dil.expr<-gcrma(Dilution,affinity.info=ai,type="affinities")
}

[Package gcrma version 2.12.1 Index]