gcrma.engine {gcrma}	R Documentation

GCRMA background adjust engine(internal function)

Description

This function adjust for non-specific binding when all arrays in the dataset share the same probe affinity information. It takes matrices of PM probe intensities, MM probe intensities, other negative control probe intensities(optional) and the associated probe affinities, and return one matrix of non-specific binding corrected PM probe intensities.

Usage


gcrma.engine(pms,mms,ncs=NULL,
                         pm.affinities=NULL,mm.affinities=NULL,anc=NULL,
                         type=c("fullmodel","affinities","mm","constant"),
                         k=6*fast+0.5*(1-fast),
                         stretch=1.15*fast+1*(1-fast),correction=1,GSB.adjust=TRUE,rho=0.7,
                         verbose=TRUE,fast=FALSE)

Arguments

pms 	 The matrix of PM intensities

mms 	 The matrix of MM intensities

ncs 	 The matrix of negative control probe intensities. When left asNULL, the MMs are considered the negative control probes.

pm.affinities 	 The vector of PM probe affinities. Note: This can be shorter than the number of rows in pms when some probes do not have sequence information provided.

mm.affinities 	 The vector of MM probe affinities.

anc 	 The vector of Negative Control probe affinities. This is ignored if MMs are used as negative controls (ncs=NULL)

type 	 "fullmodel" for sequence and MM model. "affinities" for sequence information only. "mm" for using MM without sequence information.

k 	 A tuning factor.

stretch

correction 	 .

GSB.adjust 	 Logical value. If TRUE, probe effects in specific binding will be adjusted.

rho 	 correlation coefficient of log background intensity in a pair of pm/mm probes. Default=.7

verbose 	 Logical value. If TRUE messages about the progress of the function is printed.

fast 	 Logicalvalue. If TRUE a faster add-hoc algorithm is used.

Details

Note that this expression measure is given to you in log base 2 scale. This differs from most of the other expression measure methods.

The tunning factor k will have different meainngs if one uses the fast (add-hoc) algorithm or the empirical bayes approach. See Wu et al. (2003)

Value

A matrix of PM intensties.

Author(s)

Rafeal Irizarry & Zhijin Wu

See Also

gcrma.engine2

[Package gcrma version 2.12.1 Index]