getCDF {gcrma}	R Documentation

Functions for Automatic Download of Packages

Description

These are internal functions that are called by justGCRMA and GCRMA in order to automatically download and install cdf environments and probe packages.

Usage


getCDF(cdfname, lib = .libPaths()[1], verbose = TRUE)
getProbePackage(probepackage, lib = .libPaths()[1], verbose = TRUE)

Arguments

cdfname 	 Name of the cdfenv to install.

probepackage 	 Name of the probe package to install.

lib 	 Directory of the R library where packages will be installed.

verbose 	 Output informative comments? Defaults to TRUE

Value

Nothing is returned. These functions are called simply to install environments.

Author(s)

James W. MacDonald, based on getCDFinfo, written by Jeff Gentry.

See Also

~~objects to See Also as getCDFinfo

[Package gcrma version 2.12.1 Index]