justGCRMA {gcrma}	R Documentation

Compute GCRMA Directly from CEL Files

Description

This function converts CEL files into an ExpressionSet using the robust multi-array average (RMA) expression measure with help of probe sequences.

Usage


            just.gcrma(..., filenames=character(0),
                       phenoData=new("AnnotatedDataFrame"),
                       description=NULL,
                       notes="", compress=getOption("BioC")$affy$compress.cel,
                       normalize=TRUE, bgversion=2, affinity.info=NULL,
                       type=c("fullmodel","affinities","mm","constant"),
                       k=6*fast+0.5*(1-fast), stretch=1.15*fast+1*(1-fast),
                       correction=1, rho=0.7, optical.correct=TRUE,
                       verbose=TRUE, fast=TRUE, minimum=1, optimize.by = c("speed","memory"),
                       cdfname = NULL, read.verbose = FALSE)

            justGCRMA(..., filenames=character(0),
                     widget=getOption("BioC")$affy$use.widgets,
                     compress=getOption("BioC")$affy$compress.cel,
                     celfile.path=getwd(),
                     sampleNames=NULL,
                     phenoData=NULL,
                     description=NULL,
                     notes="",
                     normalize=TRUE, 
                     bgversion=2, affinity.info=NULL,
                     type=c("fullmodel","affinities","mm","constant"),
                     k=6*fast+0.5*(1-fast), stretch=1.15*fast+1*(1-fast),
                     correction=1, rho=0.7, optical.correct=TRUE,
                     verbose=TRUE, fast=TRUE, minimum=1, optimize.by = c("speed","memory"),
                     cdfname = NULL, read.verbose = FALSE)

Arguments

... 	 file names separated by comma.

filenames 	 file names in a character vector.

widget 	 a logical specifying if widgets should be used.

compress 	 are the CEL files compressed?

phenoData 	 a AnnotatedDataFrame object.

description 	 a MIAME object.

notes 	 notes.

affinity.info 	 NULL or a list of three components: apm,amm and index, for PM probe affinities, MM probe affinities, the index of probes with known sequence, respectively.

type 	 "fullmodel" for sequence and MM model. "affinities" for sequence information only. "mm" for using MM without sequence information.

k 	 A tuning factor.

rho 	 correlation coefficient of log background intensity in a pair of pm/mm probes. Default=.7.

stretch 	 .

correction 	 .

normalize 	 Logical value. If TRUE, then normalize data using quantile normalization.

optical.correct 	 Logical value. If TRUE, then optical background correction is performed.

verbose 	 Logical value. If TRUE, then messages about the progress of the function is printed.

fast 	 Logical value. If TRUE, then a faster add-hoc algorithm is used.

optimize.by 	 "speed" will use a faster algorithm but more RAM, and "memory" will be slower, but require less RAM.

bgversion 	 integer value indicating which RMA background to use 1: use background similar to pure R rma background given in affy version 1.0 - 1.0.2 2: use background similar to pure R rma background given in affy version 1.1 and above.

minimum 	 .

celfile.path 	 a character denoting the path 'ReadAffy' should look for cel files.

sampleNames 	 a character vector of sample names to be used in the 'AffyBatch'.

cdfname 	 Used to specify the name of an alternative cdf package. If set to NULL, the usual cdf package based on Affymetrix' mappings will be used. Note that the name should not include the 'cdf' on the end, and that the corresponding probe package is also required to be installed. If either package is missing an error will result.

read.verbose 	 Logical value. If TRUE, then messages will be printed as each celfile is read in.

Details

This method should require much less RAM than the conventional method of first creating an AffyBatch and then running gcrma.

Note that this expression measure is given to you in log base 2 scale. This differs from most of the other expression measure methods.

The tuning factor k will have different meanings if one uses the fast (add-hoc) algorithm or the empirical Bayes approach. See Wu et al. (2003)

fast.bkg and mem.bkg are two internal functions.

Value

An ExpressionSet.

Author(s)

James W. MacDonald

[Package gcrma version 2.12.1 Index]